Support 32-bit RGBA-type formats
There are two main changes here.

Firstly, an adjustDeviceFormat method which, for the PiSP, munges the
32-bit RGBA formats back to 24-bit ones which are what is presented to
the hardware, though we specify a stride wide enough for the 32-it
ones.

We note whether a particular stream will need to 24-bit back to 32-bit
conversion on its output images. Images are intercepted on the way out
of the pipeline and where this flag is set, we do a simple-minded
conversion back to the initial RGBA type of output.