pipeline: raspberrypi: Add stream flags to RPi::Stream
Add a flags field to indicate stream state information in RPi::Stream.
This replaces the existing external_ and importOnly_ boolean flags.

Additionally, a new flag, RequiresMmap, is implemented to mmap the
stream buffers on allocation for the pipeline handler to use. A new
getBuffer() member function is used to retrieve the mmaped buffers.

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>